frontend: add hero and menu screens
