+++ frontend/package.json
@@ -13,9 +13,13 @@
   "dependencies": {
     "axios": "^1.9.0",
     "extendable-media-recorder": "^9.2.26",
+    "class-variance-authority": "^0.7.0",
+    "tailwind-merge": "^2.2.2",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "react-webcam": "^7.2.0"
+    "react-webcam": "^7.2.0",
+    "lucide-react": "^0.322.1",
+    "framer-motion": "^10.18.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
